Rename MatchPhrase as FullPhrase and use matchUtterance for all methods
